backend/requirements.txt
flask>=3.0.0
sentence-transformers>=2.6.1
torch>=2.1.0
requests>=2.31.0
Pillow>=10.0.0
